refactor: Update ArrowCarousel component styles for NextArrow and PrevArrow
This commit updates the styles of the NextArrow and PrevArrow components in the ArrowCarousel component. The background color has been changed from blue to a darker neutral color, and the hover background color has been changed to a lighter neutral color. These changes enhance the visual presentation and user experience of the ArrowCarousel component.