refactor: Update SVG fill colors in bg.svg and popcorn-svgrepo-com.svg
This commit updates the SVG fill colors in the bg.svg and popcorn-svgrepo-com.svg files. In bg.svg, the fill color has been changed from #551708 to #ab1406, and the fill-opacity has been adjusted to 0.6. In popcorn-svgrepo-com.svg, the fill color has been changed to #000000. These changes enhance the visual appearance of the SVG images.