refactor: Add lazy loading for TMDB genre movie and TV show data
This commit adds lazy loading functionality for the TMDB genre movie and TV show data in the `dataTMDBGenreMovie.js` and `dataTMDBGenreTvShow.js` files. It improves the performance of the pages by loading the data asynchronously, reducing the initial load time. The code now fetches the data using the `listFetch` function, which includes the API token in the request headers. These changes enhance the user experience and optimize the loading of genre movie and TV show data.